code cleanup for data collection
