+++ towhee/functional/mixins/__init__.py
@@ -21,13 +21,12 @@ from .state import StateMixin
 from .metric import MetricMixin
 from .ray import RayMixin
 from .serve import ServeMixin
-from .milvus_mixin import MilvusMixin
-from .save import SaveMixin
+from .milvus import MilvusMixin
 
 
 class DCMixins(DatasetMixin, DispatcherMixin, DisplayMixin, ParallelMixin,
-                ComputerVisionMixin, StateMixin, MetricMixin,
-                RayMixin, ServeMixin, MilvusMixin, SaveMixin):
+               ComputerVisionMixin, StateMixin, MetricMixin, RayMixin,
+               ServeMixin, MilvusMixin):
 
     def __init__(self) -> None:  # pylint: disable=useless-super-delegation
         super().__init__()
